+++ CHANGELOG.md
@@ -1,9 +1,18 @@
-## [v2.4.0 - 2025-10-20](https://github.com/thuongtruong109/flashot/releases/tag/v2.4.0)
+## [v2.5.0 - 2025-11-01](https://github.com/thuongtruong109/flashot/releases/tag/v2.5.0)
 
 ### 🚀 Features
 
 - ## **playground**:
-  - stable gutter for setting scroll bar when dropdown open - by @thuongtruong109 [<samp>()</samp>]()
+  - Custom border type (solid, dashed, dotted, double), color, width, and opacity - by @thuongtruong109 [<samp>()</samp>]()
+
+##### [View changes on GitHub](https://github.com/thuongtruong109/flashot/compare/v2.4.0...v2.5.0) - [Full Changelog](https://github.com/thuongtruong109/flashot/commits/v2.5.0)
+
+## [v2.4.0 - 2025-10-26](https://github.com/thuongtruong109/flashot/releases/tag/v2.4.0)
+
+### 🚀 Features
+
+- ## **playground**:
+  - stable gutter for setting scroll bar when dropdown open - by @thuongtruong109 [<samp>(18ab1)</samp>](https://github.com/thuongtruong109/flashot/commit/18ab1e0f70e513c845ee80deeb3f611f87f069cf)
   - Hide editor ruler when mouse out & enhance background UI - by @thuongtruong109 [<samp>(d0749)</samp>](https://github.com/thuongtruong109/flashot/commit/d07498bb5ff1e1df02a729d9c8b5b7ecddbc335e)
   - Improved manifest and social sharing support - by @thuongtruong109 [<samp>(1c382)</samp>](https://github.com/thuongtruong109/flashot/commit/1c382af38480049e77b6fee978d8e07dff1d6aea)
   - Add action buttons (clear/copy/share) - by @thuongtruong109 [<samp>(e73fd)</samp>](https://github.com/thuongtruong109/flashot/commit/e73fdb389971fae9ab7cdd8a3306a897411a1b1c)